k3s-worker-node/app/requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorzero==2.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
gpio==1.0.0
gpiozero==1.6.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.8
pigpio==1.78
pydantic==1.10.6
python-dotenv==0.21.1
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
RPi.GPIO==0.7.1
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.15.0
